+++ base/sysimg.jl
@@ -220,7 +220,7 @@ include("broadcast.jl")
 importall .Broadcast
 
 # define the real ntuple functions
-@generated function ntuple(f::F, ::Type{Val{N}}) where {F,N}
+@generated function ntuple(f::F, ::Val{N}) where {F,N}
     Core.typeassert(N, Int)
     (N >= 0) || return :(throw($(ArgumentError(string("tuple length should be ≥0, got ", N)))))
     return quote
@@ -229,7 +229,7 @@ importall .Broadcast
         @ncall $N tuple t
     end
 end
-@generated function fill_to_length(t::Tuple, val, ::Type{Val{N}}) where {N}
+@generated function fill_to_length(t::Tuple, val, ::Val{N}) where {N}
     M = length(t.parameters)
     M > N  && return :(throw($(ArgumentError("input tuple of length $M, requested $N"))))
     return quote